refactor: sidebar 컴포넌트 디렉토리 수정
- sidebar 컴포넌트는 재사용 컴포넌트가 아니기 때문에 components/로 이동
